fixed handshake for custom entity #9 added traits for tenant entity extending fixed JWTRequest for custom tenant added ability to use any tenant in dev environment added docs
